backend/requirements.txt
# FastAPI server
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4

# HTTP & Async Fetching
httpx[http2]==0.28.1
aiohttp==3.11.11
aiofiles==24.1.0

# LinkedIn Data
linkedin-api==2.2.0

# Web Scraping & Crawling
beautifulsoup4==4.12.3
lxml==5.3.0
parsel==1.9.1
trafilatura==2.0.0

# Data Enrichment
dnspython==2.7.0
email-validator==2.2.0
phonenumbers==8.13.52
python-whois==0.9.5
tldextract==5.1.3
nameparser==1.1.3

# NLP & Text Processing
spacy==3.8.4
vaderSentiment==3.3.2
keybert==0.8.5
langdetect==1.0.9
Unidecode==1.3.8
rapidfuzz==3.11.0

# Task Queue & Cache
redis==5.2.1
celery==5.4.0
apscheduler==3.11.0

# Utilities
python-dotenv==1.0.1
orjson==3.10.13
